vendetta_render: reorganize exporter
